killed walka time for wrong stats (see below)
disabled waketime because the stats where incorrect. it keeps causing a crash in vs code. I have also been using atom and sublime a lot to get around this. 

this is a big issue that i dont have time to fix until after school is done, so ill have to live with it not working for now, until I can push a fix and update it.